created a new components for the boxes in the homepage to fix use client wrong usage
